RN-977: Add option to download QR code in Meditrak (#4886)
