+++ packages/meditrak-app/app/assessment/index.js
@@ -5,6 +5,6 @@
 
 export { SurveyScreen } from './SurveyScreen';
 export { SurveysMenuScreen } from './SurveysMenuScreen';
-export { QrCodeScreen } from './qrCodeScreen';
+export { QrCodeScreen } from './QrCodeScreen';
 
 export { reducer } from './reducer';
