+++ packages/meditrak-app/package.json
@@ -57,6 +57,7 @@
     "react-native-device-info": "^7.2.1",
     "react-native-document-picker": "^8.2.1",
     "react-native-dotenv": "^0.2.0",
+    "react-native-file-access": "^1.7.1",
     "react-native-fs": "^2.16.6",
     "react-native-gesture-handler": "~1.7.0",
     "react-native-image-picker": "^2.3.4",
